backend_django/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
newspaper3k==0.2.8
python-dateutil==2.8.2
python-dotenv==1.0.0

# API Documentation
drf-yasg==1.21.7

# AI summarization (optional but recommended)
transformers>=4.36
bitsandbytes
protobuf
tokenizers
sentencepiece
accelerate